requirements.txt
streamlit
azure-ai-projects
azure-identity
